% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsAnalyst.R
\name{format_to_plot}
\alias{format_to_plot}
\title{Takes an OmicsAnalyst formatted data.frame and converts it to long format for plotting descriptive statistics and individual features}
\usage{
format_to_plot(x)
}
\arguments{
\item{x}{a data.frame}
}
\description{
Takes an OmicsAnalyst formatted data.frame and converts it to long format for plotting descriptive statistics and individual features
}
