% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsAnalyst2.R
\name{group_stats}
\alias{group_stats}
\title{Takes a plotting data.frame (typically long format) and returns the data.frame/data.table with a new column grouped by "group", transformed by "fun", and named by "name" arguments.}
\usage{
group_stats(x, group = "", variable = "", fun, name = "")
}
\arguments{
\item{x}{a data.frame}

\item{group}{column name of groups to group by}

\item{variable}{column name of variable that will be grouped by "groups"}

\item{fun}{function to apply to grouped variable}

\item{name}{desired name for new grouped variable produced by "fun"}
}
\description{
Takes a plotting data.frame (typically long format) and returns the data.frame/data.table with a new column grouped by "group", transformed by "fun", and named by "name" arguments.
}
