% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsAnalyst2.R
\name{iqr_filter}
\alias{iqr_filter}
\title{Takes an OmicsAnalyst formatted data.frame and removes varbiables(rows) with IQR below a certain threshold}
\usage{
iqr_filter(x, Threshold = 0.5)
}
\arguments{
\item{x}{a data.frame}

\item{Threshold}{a value indicating the minimum IQR to be included in the output. Varibles having IQR below this will be excluded from the output}
}
\description{
Takes an OmicsAnalyst formatted data.frame and removes varbiables(rows) with IQR below a certain threshold
}
