% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsAnalyst.R
\name{norm_Omics_df}
\alias{norm_Omics_df}
\title{Takes an OmicsAnalyst formatted data.frame and performs general logarithm transformation (glog) and/or stdandardization (autoscale) returning a normalized data.frame}
\usage{
norm_Omics_df(x, glog = T, autoscale = T)
}
\arguments{
\item{x}{a data.frame}

\item{glog}{logical argument. If TRUE (T), data will be glog normalized. If FALSE (F), data will not be glog normalized}

\item{autoscale}{logical argument. If TRUE (T), data will be autoscaled. If FALSE (F), data will not be autoscaled}
}
\description{
Takes an OmicsAnalyst formatted data.frame and performs general logarithm transformation (glog) and/or stdandardization (autoscale) returning a normalized data.frame
}
