% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsAnalyst.R
\name{plot_pca}
\alias{plot_pca}
\title{Multiple options for plotting descriptive statistics from PCA. Takes a numeric data frame as input.}
\usage{
plot_pca(
  dat,
  metadata = NULL,
  join_by_name = "Sample",
  plotting_factors_name = Group,
  plotting_factors_in = "col_names",
  x = "PC1",
  y = "PC2",
  scale = T,
  center = T,
  color = "Group",
  fill = "Group",
  plot_type = "2D",
  summarise_for_scatter = T
)
}
\arguments{
\item{dat}{A data.frame}

\item{metadata}{Optional data frame with metadata for filling, coloring, or plotting PCA results against.}

\item{join_by_name}{Name of column in metadata that matches rownames/colnames of data. Used to join metadata with data. This is usually sample names. For example, sample1.1, sample1.2, etc.}

\item{plotting_factors_name}{Name for factor you want to plot. Is made by removing two characters from the end of 'join_by_name' variable. For example, sample1.1, sample1.2 and sample2.1, sample2.2 in this variable become sample1, sample1 and sample2, sample2. Used for coloring, filling, and grouping of plot. Default name is Group}

\item{plotting_factors_in}{Location of plotting factors. Options are either 'col_names' or 'row_names'.}

\item{x}{What do you want to plot on the x-axis? Default is 'PC1'}

\item{y}{What do you want to plot on the y-axis? Default is 'PC2'}

\item{scale}{Logical. Scale data? Default is TRUE}

\item{center}{Logical. Center data? Default is TRUE}

\item{color}{Which variable to color by? Default is 'Group'.}

\item{fill}{Which variable to fill by? Default is 'Group'.}

\item{plot_type}{One of three options: '2D', 'boxplot', or 'scatter'. Default is '2D'.}

\item{summarise_for_scatter}{Logical. Plotting factors can sometimes contain psuedoreplication which inflates the p-value of this scatterplot. This option will summarize the plotting factors by mean, removing psuedoreplication for a more realistic p-value. Default is TRUE.}
}
\description{
Multiple options for plotting descriptive statistics from PCA. Takes a numeric data frame as input.
}
