% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsAnalyst2.R
\name{pre_process}
\alias{pre_process}
\title{Takes an OmicsAnalyst formatted data.frame, replaces any missing value(s) with zeros, and removes variables (rows) with constant values.}
\usage{
pre_process(x)
}
\arguments{
\item{x}{a data.frame}
}
\description{
Takes an OmicsAnalyst formatted data.frame, replaces any missing value(s) with zeros, and removes variables (rows) with constant values.
}
