% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsAnalyst.R
\name{clean_data}
\alias{clean_data}
\title{Takes an OmicsAnalyst formatted data.frame, replaces any missing value(s) with zeros, and removes varbiables(rows) with IQR below a certain threshold}
\usage{
clean_data(x, Filter = T, Threshold = 0.5)
}
\arguments{
\item{x}{a data.frame}

\item{Filter}{logical argument. If TRUE (T), data will be IQR filtered by the specified threshold. If FALSE (F), data will not be filtered}

\item{Threshold}{a value indicating the minimum IQR to be included in the output. Varibles having IQR below this will be excluded from the output}
}
\description{
Takes an OmicsAnalyst formatted data.frame, replaces any missing value(s) with zeros, and removes varbiables(rows) with IQR below a certain threshold
}
