% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsAnalyst.R
\name{expression_filter}
\alias{expression_filter}
\title{Takes a data.frame, replaces any missing value(s) with zeros, optionally performs counts per million (CPM) normalization, calculates the row-wise statistic provided by FilterFun, and filters the data with above the FilterThreshold or keeps only the top n provided by RankThreshold. The raw data is returned (not CPM normalized).}
\usage{
expression_filter(
  dat,
  CPM = TRUE,
  FilterFUN = mean,
  FilterThreshold = NULL,
  RankThreshold = NULL
)
}
\arguments{
\item{dat}{A data.frame}

\item{CPM}{Logical. Whether or not to normalize columns by CPM}

\item{FilterFUN}{Row-wise function to use for filtering (mean, max, median, IQR, etc..).}

\item{FilterThreshold}{Threshold value. Features with FilterFUN output higher than this value will be kept.}

\item{RankThreshold}{Threshold value. Features will be ranked with FilterFUN and the top n values provided by RankThreshold will be kept.}
}
\description{
Takes a data.frame, replaces any missing value(s) with zeros, optionally performs counts per million (CPM) normalization, calculates the row-wise statistic provided by FilterFun, and filters the data with above the FilterThreshold or keeps only the top n provided by RankThreshold. The raw data is returned (not CPM normalized).
}
